Move maven2 specific bits under maven2 package
Metadata merge is maven2 format specific.
